Refactor | Extract bullet generation to separate file
